Changed organization, made state machine more RP-focused, general changes
