+++ EmbeddedCode/RaspberryPi/constants.py
@@ -1,4 +1,5 @@
-#Constants used for payment calc / equivalent of #defines in c
+#Constants used for payment calc and state machine + inits
+
 #Lillian Cordelia Gwendolyn 07/14/2021 @ Wonderfil
 
 #enum not strictly required but useful for organization
@@ -45,20 +46,27 @@ WF_TAP_THREE_ADC_PIN = 4
 WF_TAP_FOUR_ADC_PIN = 6
 '''
 
+#number of taps we are using
+NUM_TAPS = 4
+
+#.5V when converted to ADCs 0-1024 scale = 155
+#number is used as the maximum for us to consider the pot to still be off
+ADC_MAX_OFF_DIGITAL_VOLTAGE = 155
+
 #enum to organize by tap numbers so i dont need to use numbers
 #doubles as port used for ADC chip reading
 #0-7 correspond to pins 0-7 on the chip
 class WF_TAP(enum.Enum):
-    NONE = -1
-    ONE = 0
-    TWO = 2
-    THREE = 4
-    FOUR = 6
+	NONE = -1
+	ONE = 0
+	TWO = 2
+	THREE = 4
+	FOUR = 6
 
 class WF_STATE(enum.Enum):
-    ERROR = -1
-    WAITING_FOR_CUSTOMER = 0
-    TAP_SELECTED = 1
-    READY_FOR_POUR = 2
-    RUNNING = 3
-    POUR_COMPLETED = 4
+	ERROR = -1
+	WAITING_FOR_CUSTOMER = 0
+	FOB_SELECTED = 1
+	READY_FOR_POUR = 2
+	RUNNING = 3
+	POUR_COMPLETED = 4
